added landing page & routing
