+++ package.json
@@ -11,6 +11,7 @@
   "devDependencies": {
     "@sveltejs/vite-plugin-svelte": "^6.1.1",
     "svelte": "^5.38.1",
+    "svelte-routing": "^2.13.0",
     "vite": "^7.1.2",
     "vite-plugin-singlefile": "^2.3.0"
   },
