+++ src/app.css
@@ -28,3 +28,12 @@ body {
     font-family: Avenir, Helvetica, Arial, sans-serif;
     color: #cfcfcf;
   }
+
+  .glass-container {
+    background: rgba(30, 30, 40, 0.35);
+    box-shadow: 0 8px 32px 0 rgba(31, 38, 135, 0.37);
+    backdrop-filter: blur(12px) saturate(180%);
+    -webkit-backdrop-filter: blur(12px) saturate(180%);
+    border: 1px solid rgba(255, 255, 255, 0.18);
+    border-radius: 16px;
+  }
